+++ lib/ModelMixins/ReferenceMixin.ts
@@ -106,6 +106,13 @@ function ReferenceMixin<T extends Constructor<Model<ReferenceTraits>>>(
       );
 
       if (!result.error && this.target) {
+        // Copy knownContainerUniqueIds to target
+        this.knownContainerUniqueIds.forEach(id =>
+          !this.target!.knownContainerUniqueIds.includes(id)
+            ? this.target!.knownContainerUniqueIds.push(id)
+            : null
+        );
+
         applyItemProperties(this, this.target);
       }